reduced unsigned sizes where possible - nodes and qubits down to u16

+++ src/node.rs
@@ -16,20 +16,20 @@ pub enum NodeType {
 #[derive(Debug, Clone, PartialEq)]
 pub struct Node {
     pub node_type: NodeType,
-    pub id: usize,
+    pub id: u16,
     pub label: String,
-    pub paired_data_id: Option<usize>,
+    pub paired_data_id: Option<u16>,
     pub pos: (f32, f32),
     pub busy_count: i32,
     pub cultivation_time: i32,
-    pub nbors: Vec<usize>,
+    pub nbors: Vec<u16>,
 }
 
 static USE_MAGIC_ROUTING: AtomicBool = AtomicBool::new(true);
 
 impl Node {
     /// Creates a new node with the given properties and empty neighbor set.
-    pub fn new(id: usize, paired_data_id: Option<usize>, label: String, x: f32, y: f32,
+    pub fn new(id: u16, paired_data_id: Option<u16>, label: String, x: f32, y: f32,
                node_type: NodeType, busy_count: i32, cultivation_time: i32)
                -> Self {
         Node { node_type,
@@ -48,7 +48,7 @@ impl Node {
     }
 
     /// Adds a neighbor to this node's connectivity list.
-    pub fn add_neighbor(&mut self, other: usize) {
+    pub fn add_neighbor(&mut self, other: u16) {
         if !self.nbors.contains(&other) {
             self.nbors.push(other);
         }
